+++ Readme.md
@@ -2,31 +2,30 @@
 
 ## About
 
-When it comes to modernizing the applications, migrating to Mapstruct from Dozer mapper is a primary part for most of the Java enterprise applications that are running sucessfully for more than a decade.
+When it comes to modernizing the applications, migrating to Mapstruct from Dozer mapper is a primary part for most of the Java enterprise applications that are running successfully for more than a decade.
 
 This Intellij plugin will help us to do the migration faster by generating mapstruct code for the dozer xml.
 
 ## Steps to add this plugin to ur ide
 
-Click Ctrl+Alt+s shortcut on Intellij IDE to go plugins.
-Go to install from disk options.
-Import the downloaded  jar from disk. <br>
-![to](Documents/to.png "to")
-
 ## How to Use The Plugin
 - Select the xml code.
 - press Ctrl + Shift + ` shortcut  and provide the details.
-- Code will be generated at the specified location.<br>
+- Code will be generated at the specified location.
+- - Map-id at field levels are added.<br>
 
 ![imageCheckBox](Documents/imageCheckBox.png "imageCheckBox")
 - Uncheck the checkBok to avoid writing and just view the generated code in pop up.
 - Go to action by pressing (⌘+Shift+A or Ctrl+Shift+A)  and change the shortcut by yourself.
 
-
+Alternate, 
+- Just right click on selected text click on *GenerateMapping*  which appears first on popup menu.<br>
 
 ![menu](Documents/menu.png "menu")
-- Just right click on selected text click on *GenerateMapping*  which appears first on popup menu.
-- Map-id at field levels are added.
+<br>
+
+![view](Documents/view.png "view")
+
